refactor(api): split company-review repository into add and get files
